requirements.txt
trakt.py
schedule
cinemagoer==2022.2.11